DodgeBall game
You are a paddle on one side of the bank of 7-segment display.
The enemy is the "ball" that gets thrown at you from the other side.
Dodge it by pressing up or down to move your paddle.
Hit both button to start/reset the game.

This is the VHDL files only, except for a Makefile and some support
files to build via ghdl. It won't run in GHDL, though, and it needs
to be integrated into the build environment for some FPGA.

My FPGA is the Terasic DE10 MAX Lite (or something like that), which is
a 10M50DAF484C7G FPGA. It looks like this:
https://www.terasic.com.tw/attachment/archive/1021/image/45degree.jpg

The DodgeBall.v file is the integration into Quartus via the Terasic
System Builder. There are a bunch of other files needed, but if you're
really interested, you can probably work it out.